+++ CHANGELOG.md
@@ -14,6 +14,17 @@
 ### Fixed
 ### Security
 
+
+
+## [1.3.0] - 2026-06-28
+
+### Changed
+- あまり表に出ないクラス名、関数名などを変更。
+- setAutoDirection()の機能をActionObjectからObjectA3に移動しました
+### Fixed
+- SimplePhysicsモードに変更する前のscale設定がColliderDescに
+  反映されるようになった。
+
 ## [1.2.0] - 2026-06-15
 
 ### Changed